Split contents of interfaces.ts
