+++ src/constants.ts
@@ -1,0 +1,5 @@
+export enum TimerType {
+  once = 0,
+  periodic = 1,
+  cron = 2,
+}
