+++ src/itiming_rule.ts
@@ -1,0 +1,9 @@
+export interface ITimingRule {
+  second?: number;
+  minute?: number;
+  hour?: number;
+  date?: number;
+  month?: number;
+  year?: number;
+  dayOfWeek?: number;
+}
